Issue 34: Move MemberBundle into WebBundle
