--- src/Khatovar/Bundle/MemberBundle/DependencyInjection/KhatovarMemberExtension.php
@@ -1,43 +1,0 @@
-<?php
-
-/**
- * This file is part of KhatovarWeb.
- *
- * Copyright (c) 2015 Damien Carcel (https://github.com/damien-carcel)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-namespace Khatovar\Bundle\MemberBundle\DependencyInjection;
-
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-
-/**
- * @author Damien Carcel <damien.carcel@gmail.com>
- */
-class KhatovarMemberExtension extends Extension
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function load(array $configs, ContainerBuilder $container)
-    {
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('form_subscribers.yml');
-        $loader->load('form_types.yml');
-    }
-}
