+++ src/Khatovar/Bundle/WebBundle/Entity/MemberRepository.php
@@ -19,7 +19,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Khatovar\Bundle\MemberBundle\Entity;
+namespace Khatovar\Bundle\WebBundle\Entity;
 
 use Doctrine\ORM\EntityRepository;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
