+++ src/Khatovar/Bundle/WebBundle/Form/Subscriber/AddPortraitSubscriber.php
@@ -19,7 +19,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Khatovar\Bundle\MemberBundle\Form\Subscriber;
+namespace Khatovar\Bundle\WebBundle\Form\Subscriber;
 
 use Doctrine\ORM\EntityRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
