+++ src/Khatovar/Bundle/WebBundle/Resources/config/routing.yml
@@ -1,10 +1,10 @@
-khatovar_web_camp:
-    path:     /camp
-    defaults: { _controller: KhatovarWebBundle:Web:camp }
-
 khatovar_appearance:
     resource: '@KhatovarWebBundle/Resources/config/routing/appearance.yml'
     prefix:   /prestations
+
+khatovar_camp:
+    path:     /camp
+    defaults: { _controller: KhatovarWebBundle:Web:camp }
 
 khatovar_contact:
     resource: '@KhatovarWebBundle/Resources/config/routing/contact.yml'
@@ -17,3 +17,7 @@ khatovar_exaction:
 khatovar_homepage:
     resource: '@KhatovarWebBundle/Resources/config/routing/homepage.yml'
     prefix:   /accueil
+
+khatovar_member:
+    resource: '@KhatovarWebBundle/Resources/config/members.yml'
+    prefix:   /membres
