+++ src/Khatovar/Bundle/WebBundle/Resources/views/Member/edit.html.twig
@@ -8,5 +8,5 @@
 
     <h1>Édition de la page de membre</h1>
 
-    {% include 'KhatovarMemberBundle:Member:_form.html.twig' with {'form': edit_form} %}
+    {% include 'KhatovarWebBundle:Member:_form.html.twig' with {'form': edit_form} %}
 {% endblock %}
