+++ src/Khatovar/Bundle/WebBundle/Resources/views/Member/new.html.twig
@@ -8,5 +8,5 @@
 
     <h1>Création d'une nouvelle page de membre</h1>
 
-    {% include 'KhatovarMemberBundle:Member:_form.html.twig' with {'form': form} %}
+    {% include 'KhatovarWebBundle:Member:_form.html.twig' with {'form': form} %}
 {% endblock %}
